Bluetooth: Fix for not to trigger Epilog process if no interface
Fix for not to trigger epilog process when SoC is not initialized of
failed to get initialized properly. Clean up process tries to trigger
epliog process if the worker thread is running. If the SoC is not
initialized properly, then there will be a chances of hci reset command
not being sent SoC and the stack waits for response for SoC for 3 sec.
Cleanup is delayed by 3 sec causing issues with BT on/off

Change-Id: I40bf2fbf33f8b5ae3ac52b69f219ac6db02d2591
CRs-fixed: 676376